separate deploy and start for testing
